Added CI helper scripts to skip unneeded builds (#2)
To avoid performing unneccessary builds we have a helper script
ci-checkchanges.sh which is used in VSTS to check if there are
any actual changes in this changeset that affect the webapp and
set a variable for VSTS.

We also configure VSTS to check that variable and skip actual build
steps if the build is not needed.

Similarly there is ci-checkchanges.cmd that does exactly the same
thing for builds that run on Windows.

I also added ci-build.sh script which includes all shell commands
used in webapp build. Having it under source control is much nicer than
just saving these same commands in VSTS (because we can now enjoy
the benefits of version control for build commands too).