Introduced shared code directory for mobile app
Summary
=======

In preparation to have a Client App in addition to the existing
Stylist App I added the support for having code that is common
for both mobile apps.

The proposed directory structure is this:

\mobile
  \client  <-- Client App (Ionic)
  \stylist <-- Stylist App (Ionic)
  \shared  <-- shared code for both apps

When the Client App is actually created some of the files
that are currently in mobile/stylist directory need to be moved
to mobile/shared directory (e.g. some API services, models, etc.)

Solution
========

To support this structure the following is done:

1. In tsconfig.json add baseUrl and paths to map. This
adds TS compile support.

2. In webpack.config.js add support for @shared alias
to map to the same directory. This adds webpack bundling
support.

3. In watch.config add files "shared" directory to the
the watched list. This enables auto-recompiled in livereload
mode for shared files.

4. Move all files that previously were in "mobile" directory
to "mobile/stylist" directory.

5. Add example files in shared directory and demonstrate
that they can be used in the Stylist App (used in app.component.ts).

6. I had to disable "no-implicit-dependencies" linting rules,
I was unable to make it work with the shared directory properly
(Known issue: https://github.com/palantir/tslint/issues/3364).

7. By popular demand used the same aliasing technique to
make ~ an alias for root source code directory. Now we can import
files like this:
    import { TodayComponent } from '~/today/today.component';

TODO
====

1. Verify that we can also create tests for shared code.

2. Split ios-cert and have separate certificates for Stylist
and Client apps

3. Add separate CI build for Client App when it is created
and split c-checkchanges script to check corresponding directory
for each mobile apps plus the shared directory.