Add API to pre-fetch to-be created appointment details
Allow stylist to preview the appointment for a client before appointment
is actually created; this may be required to preview duration, actual
price that client will see (with or without discounts) and conflicting
appointments (if any).

If client is not in the system, client's price will always  be equal to
the base price of the service.

Github issue: #105