Expand list of appointments for Today's screen endpoint
Previously, for Today's screen API was returning only current and (if
exist) - next appointment. This commit extends this list to include
more appointment types, namely - all appointments for the day, excluding:
- appointments cancelled by the stylist themselves;
- appointments successfully paid (checked out)
- appointments marked by the stylist as 'No Show'.
Appointments cancelled by the client will still be shown (so that
stylist is aware of what is happening).

Apart from including more types, appointments in the list will be
fetched from the beginning of the day, not from the current moment.

Github issue: # 113