Fix error messages on Logout
The bug was happening because logout was performed from a popover
component and it was attempting to set root page to first screen using
the NavController that is available in popover.

In popovers the NavController is a different instance
and attempting to call NavController.setRoot() to
return to FirstScreen will result in misterious and
spurious errors.
See solution here: https://github.com/ionic-team/ionic/issues/8437#issuecomment-260375966

Trello: https://trello.com/c/qn3oJW8z/24-fix-loading-stylist-summary-faild