[HOTFIX] Fix price calculation affected by appointment early creation
Presently when an appointment is created, it's effective price can be
different from what a client/stylist had seen in the price preview. This
happens due to the fact that these two APIs are decoupled, and when an
appointment is created, price is recalculated **after** the appointment
is added. This adds to the day's demand, hence price may be different
from the previous request.

This commit fixes the situation by pre-calclulating prices and fixing
them **before** the appointment is actually created. Apart from that,
this commit truncates calculated prices to provide whole values.

Github issue: #