Disable First Time Discount since it makes no sense without Client App
Currently this discount is applied for any client for which the stylist
books an appointment for the first time using Made Pro. However in
reality this may not be a new client, they may be an old client of the
stylist, but Made Pro just has no way of knowing it. Applying the first
time discount to these clients makes no sense: stylist wins nothing and
we win nothing.

The First Time Discount will make sense for stylist only if it is
really a new client and will make sense for us only if it results in
the client downloading and starting to use our Client App (which
doesn't exist yet).

We disable this discount for now. We will re-enable it as soon as the
Client App is ready and the discount will be applied only to new
clients who perform the first booking using the App (but not to first
booking if it is done by stylist using Stylist App).

Github issue: #375