Auto-create superuser on staging and production deploy
Management command to create (or update) superuser in non-interactive mode,
with password set in the environment variable. We need to be able
to automatically create a superuser for staging and production, yet
standard Django's createsuperuser command is not capable of running in
non-interactive mode, nor it is able to amend existing superuser's password.
This command aims to solve this. If superuser is missing (e.g. after
complete DB rebuilding) the command will create a superuser with
pre-defined email and password taken from an envvariable.

If superuser already exists, command will just update superuser's
password.