[HOTFIX] Refactor client details API to accept UUID of client
Previously, stylists API was primarily operating with UUIDs of
ClientOfStylist objects. With the recent changes to search and list
APIs and changing criteria of access to client's details from being
based on the presence of ClientOfStylist relation to presence of
PreferredStylist relation, Stylist's API now operates with UUIDs
of Client object. This fix refactors client details API to use UUID
of the Client object, and is the entry code for the next PR of the
endpoint for retrieving client's calendar object (which is also UUID-
based).