[HOTFIX] Fix reporting to slack logic for Twilio messages
There is a bug in function that sends messages via twilio: if a
message was actually sent through Twilio, it would never be reported
in slack. Although this is a desired behavior, it's still a bug, because
this logic must be controlled by external settings rather than a bug
(in our crease - premature return from function) in the code.