[HOTFIX] Fix price calculation for zero prices
The problem was there long ago, but it started to manifest itself just
recently when we turned off adding tax by default. But the actual problem
is that we were not defining an accumulator variable as Decimal; so
under certain circumstances we were ending up with int instead of Decimal
which was causing the exception.